backend/agent_llm/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alex-agent-llm"
version = "0.1.0"
description = "Shared OpenRouter wiring for Alex OpenAI Agents SDK services"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["alex_agent_llm"]
